ASR Agent program with Google and PocketSphinx backends (#25)
This PR adds an agent program that performs automatic speech recognition from a microphone stream. It currently supports the free PocketSphinx engine and the paid Google Cloud Speech Recognition API. The default behaviour is to print output to stdout, but it also has support for publishing output to an MQTT message bus.